+++ tests/testthat/test-subset-replace.R
@@ -19,28 +19,21 @@ test_that("One-argument [", {
 })
 
 
+test_that("Zero-argument [<-", {
+  ht <- hux(a = 1:2, b = 1:2, add_colnames = TRUE)
+  expect_silent(ht[] <- matrix(1:6, 3, 2))
+  expect_identical(ht[[3, 2]], 6L)
+
+  ht <- hux(a = 1:2, b = 1:2, add_colnames = TRUE)
+  expect_silent(ht[ , ] <- matrix(1:6, 3, 2))
+  expect_identical(ht[[3, 2]], 6L)
+})
+
+
 test_that("Subsetting preserves rownames", {
   ht <- huxtable(a = 1:3, b = 1:3)
   rownames(ht) <- letters[1:3]
   expect_equal(rownames(ht[1:2, ]), letters[1:2])
-})
-
-
-
-test_that("Subsetting cuts rowspan and colspan", {
-  ht <- hux(a = 1:3, b = 1:3, d = 1:3)
-  rowspan(ht)[1, 1] <- 3
-  colspan(ht)[1, 2] <- 2
-  ss <- ht[1:2, 1:2]
-  expect_equivalent(rowspan(ss)[1, 1], 2)
-  expect_equivalent(colspan(ss)[1, 2], 1)
-})
-
-
-test_that("Subsetting works with multirow/multicolumn cells", {
-  ht <- hux(a = 1:3, b = 1:3)
-  rowspan(ht)[1, 1] <- 2
-  expect_silent(ht[c(1, 3), ])
 })